Use 'cordova' build tag instead of runtime detection
