+++ webclient/init_www.go
@@ -1,0 +1,12 @@
+// +build !cordova
+
+package main
+
+import (
+	"sync"
+)
+
+func initCordova(_ *sync.WaitGroup) {
+	// nothing to do here
+	return
+}
